Log taskrunner start/stop/end and cleanup-skipping
Log when TaskRunner starts, stops or fails.
Also log when cleanup get's skipped.

This should remove the possible confusion that sometimes,
when taskrunner is executed with clean=never or in similar
way, it's not clear from the output if execution really
finished completely or if it died somewhere (unless
the Tasks executed do log their end on their own).